@@ -1,16 +1,16 @@
-apply plugin: 'com.android.application'
+plugins {
+    id 'com.android.application'
+}
 
 android {
-
-    compileSdkVersion 29
-    buildToolsVersion "29.0.1"
+    namespace 'com.sdk.coolfar_sdk'
+    compileSdk 34
     defaultConfig {
         applicationId "com.sdk.coolfar_sdk"
-        minSdkVersion 18
-        targetSdkVersion 29
+        minSdkVersion 21
+        targetSdkVersion 33
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -28,11 +28,11 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-//    implementation 'androidx.appcompat:appcompat:1.0.2'
-//    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test:runner:1.2.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    api project(':keepalive')
+    implementation(platform("org.jetbrains.kotlin:kotlin-bom:1.8.22"))
+    api project(':keepalive')//    implementation 'androidx.appcompat:appcompat:1.0.2'
+////    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+//    testImplementation 'junit:junit:4.13.2'
+//    androidTestImplementation 'androidx.test:runner:1.6.1'
+//    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
 //    compile(name:'keepalive-1.1.3', ext:'aar')
 }
