@@ -1,29 +1,24 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
-group='shihoo.wang'
-
-//aar名字
-def SDK_NAME = "keepalive-1.1.5.aar"
-
-//去除多余类配置
-configurations.maybeCreate("exclude")
-artifacts.add("exclude", file(SDK_NAME))
-//先生成aar后再加上去除配置
-apply from: "${project.projectDir.absoluteFile}\\excludeAar.gradle"
+//apply plugin: 'com.github.dcendents.android-maven'
+//group='shihoo.wang'
+//
+////aar名字
+//def SDK_NAME = "keepalive-1.1.5.aar"
+//
+////去除多余类配置
+//configurations.maybeCreate("exclude")
+//artifacts.add("exclude", file(SDK_NAME))
+////先生成aar后再加上去除配置
+//apply from: "${project.projectDir.absoluteFile}\\excludeAar.gradle"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "29.0.1"
-
-
+    namespace 'com.sdk.keepbackground'
+    compileSdk 34
     defaultConfig {
-        minSdkVersion 18
-        targetSdkVersion 26
+        minSdkVersion 21
+        targetSdkVersion 33
         versionCode 2
         versionName "1.1"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
     }
 
     buildTypes {
@@ -38,23 +33,27 @@ android {
         }
     }
 
-//可根据自己需求更改-+.693  gradlew againMakeJar
-    task againMakeJar(type: Copy) {
-        delete 'mylib/keeplive.aar' //删除之前的旧jar包
-        from('build/outputs/aar/') //从这个目录下取出默认jar包
-        into('/') //将jar包输出到指定目录下
-        include('keepalive-release.aar')
-        rename('keepalive-release.aar', SDK_NAME) //自定义jar包的名字
-    }
-    againMakeJar.dependsOn(build)
-}
-ext.rxJavaVersion = '2.1.2'
-ext.rxAndroidVersion = '2.0.1'
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    testImplementation 'junit:junit:4.12'
-    api 'com.android.support:appcompat-v7:26.1.0'
+////可根据自己需求更改-+.693  gradlew againMakeJar
+//    task againMakeJar(type: Copy) {
+//        delete 'mylib/keeplive.aar' //删除之前的旧jar包
+//        from('build/outputs/aar/') //从这个目录下取出默认jar包
+//        into('/') //将jar包输出到指定目录下
+//        include('keepalive-release.aar')
+//        rename('keepalive-release.aar', SDK_NAME) //自定义jar包的名字
+//    }
+//    againMakeJar.dependsOn(build)
+//}
+
+
+    ext.rxJavaVersion = '2.1.2'
+    ext.rxAndroidVersion = '2.0.1'
+
+
+    dependencies {
+        implementation fileTree(dir: 'libs', include: ['*.jar'])
+        implementation 'androidx.appcompat:appcompat:1.7.0'
 
 //    implementation "io.reactivex.rxjava2:rxjava:${rxJavaVersion}"
 //    implementation "io.reactivex.rxjava2:rxandroid:${rxAndroidVersion}"
+    }
 }
