@@ -10,9 +10,7 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.Settings;
-import android.support.annotation.RequiresApi;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
+
 import android.util.Log;
 
 
@@ -23,6 +21,10 @@ import com.sdk.keepbackground.watch.AbsServiceConnection;
 
 import static android.content.pm.PackageManager.PERMISSION_GRANTED;
 import static com.sdk.keepbackground.work.IntentWrapper.getApplicationName;
+
+import androidx.annotation.RequiresApi;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 
 
 /**
